remove ShareesController, extend ShareDialogView
